[feat/#54] 오늘의 공부 결과 화면 구현

+++ core/data/daily-study/daily-study-api/src/main/java/com/qriz/app/core/data/daily_study/daily_study_api/model/WeeklyReviewResult.kt
@@ -1,23 +1,23 @@
 package com.qriz.app.core.data.daily_study.daily_study_api.model
 
 data class WeeklyReviewResult(
-    val totalScore: Float,
+    val totalScore: Int,
     val subjectItems: List<SubjectResult>,
 )
 
 data class SubjectResult(
     val subjectName: String,
-    val score: Float,
+    val score: Int,
     val categoryItems: List<CategoryResult>
 )
 
 data class CategoryResult(
     val categoryName: String,
-    val score: Float,
+    val score: Int,
     val conceptItems: List<ConceptResult>
 )
 
 data class ConceptResult(
     val conceptName: String,
-    val score: Float,
+    val score: Int,
 )
